clean up constants and imports
